Add files via upload
GUI can be used as a temporal, visual status of the chessboard. After one has taken a look at the GUI, one needs to close it to continue
choosing a figure and calling the destination. After the move, the GUI is displayed again and the cycle repeats.

I need to find a way to get the GUI running parallel to the console.