Add files via upload
The multithreading works out! After hitting the start button player one can select with a left mouse click
on the specific white figure which figure to move. After that player one needs to right click on the destination field.
By hitting the submit the field gets visually update. Now its the turn of Player black and the cycle continues.